+++ build.gradle
@@ -4,7 +4,7 @@ plugins {
 
 }
 
-group 'ru.gosteva'
+group 'com.github.egosteva'
 version '1.0-SNAPSHOT'
 
 repositories {
