+++ src/main/java/com/github/egosteva/Main.java
@@ -1,4 +1,4 @@
-package ru.gosteva;
+package com.github.egosteva;
 
 public class Main {
     public static void main(String[] args) {
