+++ src/test/java/com/github/egosteva/simple/NegativeTests.java
@@ -1,22 +1,30 @@
-package simple;
+package com.github.egosteva.simple;
 
-import org.junit.jupiter.api.Disabled;
 import org.junit.jupiter.api.Tag;
 import org.junit.jupiter.api.Test;
 
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
 @Tag("simple")
-public class SkippedTests {
+public class NegativeTests {
     @Test
-    @Disabled
     void test1() {
         assertTrue(false);
     }
 
     @Test
-    @Disabled("Some reason")
     void test2() {
         assertTrue(false);
     }
+
+    @Test
+    void test3() {
+        assertTrue(false);
+    }
+
+    @Test
+    void test4() {
+        assertTrue(false);
+    }
+
 }
