+++ src/test/java/com/github/egosteva/simple/PositiveTests.java
@@ -1,4 +1,4 @@
-package simple;
+package com.github.egosteva.simple;
 
 import org.junit.jupiter.api.Tag;
 import org.junit.jupiter.api.Test;
