+++ src/test/java/com/github/egosteva/simple/SkippedTests.java
@@ -1,30 +1,22 @@
-package simple;
+package com.github.egosteva.simple;
 
+import org.junit.jupiter.api.Disabled;
 import org.junit.jupiter.api.Tag;
 import org.junit.jupiter.api.Test;
 
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
 @Tag("simple")
-public class NegativeTests {
+public class SkippedTests {
     @Test
+    @Disabled
     void test1() {
         assertTrue(false);
     }
 
     @Test
+    @Disabled("Some reason")
     void test2() {
         assertTrue(false);
     }
-
-    @Test
-    void test3() {
-        assertTrue(false);
-    }
-
-    @Test
-    void test4() {
-        assertTrue(false);
-    }
-
 }
